The app is completed with every feature desired currently.

+++ app/src/main/java/com/example/loginactivity/WelcomeStudent.java
@@ -1,17 +1,20 @@
 package com.example.loginactivity;
 
 import androidx.appcompat.app.AppCompatActivity;
-
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.TextView;
 
 public class WelcomeStudent extends AppCompatActivity {
-    EditText text;
+    TextView text;
     Button logout;
     Button ratelecture;
     SharedPreferences sharedPreferences;
@@ -19,12 +22,11 @@ public class WelcomeStudent extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_welcome_student);
-        text=(EditText)findViewById(R.id.student_welcome);
-        logout=(Button)findViewById(R.id.logout_button_student);
+        text=(TextView)findViewById(R.id.student_welcome);
         ratelecture=(Button)findViewById(R.id.button_post_rating);
         sharedPreferences = getSharedPreferences("studentdetails", Context.MODE_PRIVATE);
         String username=sharedPreferences.getString("username","");
-        text.setText("                       Welcome   "+username);
+        text.setText("Welcome "+username);
         ratelecture.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -32,26 +34,35 @@ public class WelcomeStudent extends AppCompatActivity {
                 startActivity(intent);
             }
         });
-        logout.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                SharedPreferences sharedpreferences = getSharedPreferences("studentdetails", Context.MODE_PRIVATE);
-                SharedPreferences.Editor editor = sharedpreferences.edit();
-                editor.clear();
-                editor.apply();
-                Intent welcomeintent;
-                welcomeintent = new Intent(com.example.loginactivity.WelcomeStudent.this,com.example.loginactivity.MainMenuActivity.class);
-                welcomeintent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-                startActivity(welcomeintent);
-                finish();
-            }
-        });
-
-
-
     }
 
     @Override
     public void onBackPressed(){
+    }
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu){
+
+        MenuInflater inflater=getMenuInflater();
+        inflater.inflate(R.menu.mainmenu,menu);
+        return true;
+    }
+    public boolean onOptionsItemSelected(MenuItem item){
+        if(item.getItemId()==R.id.item_aboutapp){ Intent welcomeintent;
+            welcomeintent = new Intent(com.example.loginactivity.WelcomeStudent.this,com.example.loginactivity.abouttheapp.class);
+            startActivity(welcomeintent);
+        }
+        if(item.getItemId()==R.id.item_logout){
+            SharedPreferences sharedpreferences = getSharedPreferences("studentdetails", Context.MODE_PRIVATE);
+            SharedPreferences.Editor editor = sharedpreferences.edit();
+            editor.clear();
+            editor.apply();
+            Intent welcomeintent;
+            welcomeintent = new Intent(com.example.loginactivity.WelcomeStudent.this,com.example.loginactivity.MainMenuActivity.class);
+            welcomeintent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+            startActivity(welcomeintent);
+            finish();
+        }
+        return super.onOptionsItemSelected(item);
+
     }
 }
